requirements.txt
altair==4.1.0
folium==0.12.1.post1
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.4
Pillow==9.1.0
pyrsistent==0.17.3
scikit_learn==1.0.2
seaborn==0.11.1
sympy==1.8
